cmd/ore/aws: Merge create_image.go into upload.go
No textual changes.
